fix(zig): match python simulation across a bot replay sweep
A sweep of bot-played replays (every quest level, survival, rush, typo,
several seeds) now verifies identically in both ports.

- creatures: ping-pong heading jitter, AI link-death damage path, spawn slot
  release on corpse cull, grid formation targets, template tints
- quest spawns: ring/radial/spiral positions truncated from f64
- rocket minigun: single muzzle sprite, game-tune draw order, trail math,
  burst scale, impulse, detonation state, secondary hit counts
- player damage: exact-zero health is lethal unless preserve_bugs
- bonuses: pending freeze/fireblast effects before pickup fx
- f64 stat, pellet and damage-pool math where python computes in doubles
- dbg traces carry typo/tutorial checkpoint snapshots
- tests: bot regression replays, full-trace diffs, exact spawn tables for all
  levels, opt-in sweep (CRIMSON_ZIG_BOT_SWEEP=1)

+++ crimson-zig/src/quest_spawn/logic_tier4.zig
@@ -316,7 +316,7 @@ fn build_407_gauntlet(
 
     const ring_count = player_count + 9;
     if (ring_count > 0) {
-        const step = std.math.tau / @as(f32, @floatFromInt(ring_count));
+        const step = std.math.tau / @as(f64, @floatFromInt(ring_count));
         try common.appendRingSpawns(
             out_entries,
             len,
@@ -352,7 +352,7 @@ fn build_407_gauntlet(
 
     const outer_count = player_count + 0x11;
     if (outer_count > 0) {
-        const step = std.math.tau / @as(f32, @floatFromInt(outer_count));
+        const step = std.math.tau / @as(f64, @floatFromInt(outer_count));
         try common.appendRingSpawns(
             out_entries,
             len,
@@ -592,7 +592,7 @@ fn build_410_the_end_of_all(
         var trigger: i32 = 62_800;
         var ring_index: i32 = 0;
         while (ring_index < 12) : (ring_index += 1) {
-            const angle = (@as(f32, @floatFromInt(ring_index)) + 1.0) * 0.5235988;
+            const angle = @as(f64, @floatFromInt(ring_index + 1)) * 0.5235988;
             try common.appendSpawn(
                 out_entries,
                 len,
