+++ crimson-zig/src/runtime/fire_recipes.zig
@@ -7,26 +7,28 @@ const weapon_data = @import("weapon_data.zig");
 pub const WeaponId = game_ids.WeaponId;
 pub const ProjectileTypeId = game_ids.ProjectileTypeId;
 
+/// Pellet jitter and speed-scale rules are evaluated in double precision and
+/// rounded once when stored in the f32 projectile fields.
 pub const PelletJitterRule = union(enum) {
     none,
     modulo_centered: struct {
         modulo: u32,
         center: i32,
-        step: f32,
+        step: f64,
     },
     mask_centered: struct {
         mask: u32,
         center: i32,
-        step: f32,
+        step: f64,
     },
 };
 
 pub const SpeedScaleRule = union(enum) {
     none,
     modulo: struct {
-        base: f32,
+        base: f64,
         modulo: u32,
-        step: f32,
+        step: f64,
     },
 };
 
@@ -89,7 +91,7 @@ const gauss_ion_speed_scale: SpeedScaleRule = .{
     },
 };
 
-pub fn pelletJitterStepForWeapon(weapon_id: WeaponId) f32 {
+pub fn pelletJitterStepForWeapon(weapon_id: WeaponId) f64 {
     return switch (weapon_id) {
         .shotgun, .jackhammer => 0.0013,
         .sawed_off_shotgun => 0.004,
